dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\dev-requirements.in'
#
backcall==0.1.0           # via ipython
colorama==0.4.1           # via ipython
decorator==4.4.0          # via ipython, traitlets
ipykernel==5.1.1
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0            # via ipykernel
jedi==0.14.1              # via ipython
jupyter-client==5.3.1     # via ipykernel
jupyter-core==4.5.0       # via jupyter-client
parso==0.5.1              # via jedi
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.9     # via ipython
pygments==2.4.2           # via ipython
python-dateutil==2.8.0    # via jupyter-client
pyyaml==5.1.2
pyzmq==18.0.2             # via jupyter-client
six==1.12.0               # via prompt-toolkit, python-dateutil, traitlets
tornado==6.0.3            # via ipykernel, jupyter-client
traitlets==4.3.2          # via ipykernel, ipython, jupyter-client, jupyter-core
wcwidth==0.1.7            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via ipython
